pkg/seq/ast/int/lexer: moves the package inside subdir of ast package
