+++ pkg/sequence/ast/internal/lexer/from.go
