+++ pkg/sequence/ast/internal/lexer/from_test.go
