+++ pkg/sequence/ast/internal/lexer/nodes.go
@@ -1,6 +1,6 @@
 package lexer
 
-import "github.com/ufukty/diagramer/pkg/sequence/lexer/tokens"
+import "github.com/ufukty/diagramer/pkg/sequence/ast/internal/lexer/tokens"
 
 // MARK: Lifelines and rel.
