+++ pkg/sequence/ast/internal/lexer/tokens/tokens.go
